Board 클래스 리펙터링
점점 복잡해지고 거대해지는 Board클래스를 리펙터링함
View, Board, Game(Rule)로 역할을 나누고 클래스 분리
